libnetwork/drivers/bridge: remove redundant type conversion

+++ libnetwork/drivers/bridge/port_mapping_linux_test.go
@@ -40,9 +40,9 @@ func TestPortMappingConfig(t *testing.T) {
 		t.Fatalf("Failed to setup driver config: %v", err)
 	}
 
-	binding1 := types.PortBinding{Proto: types.SCTP, Port: uint16(300), HostPort: uint16(65000)}
-	binding2 := types.PortBinding{Proto: types.UDP, Port: uint16(400), HostPort: uint16(54000)}
-	binding3 := types.PortBinding{Proto: types.TCP, Port: uint16(500), HostPort: uint16(65000)}
+	binding1 := types.PortBinding{Proto: types.SCTP, Port: 300, HostPort: 65000}
+	binding2 := types.PortBinding{Proto: types.UDP, Port: 400, HostPort: 54000}
+	binding3 := types.PortBinding{Proto: types.TCP, Port: 500, HostPort: 65000}
 	portBindings := []types.PortBinding{binding1, binding2, binding3}
 
 	sbOptions := make(map[string]interface{})
@@ -126,9 +126,9 @@ func TestPortMappingV6Config(t *testing.T) {
 	}
 
 	portBindings := []types.PortBinding{
-		{Proto: types.UDP, Port: uint16(400), HostPort: uint16(54000)},
-		{Proto: types.TCP, Port: uint16(500), HostPort: uint16(65000)},
-		{Proto: types.SCTP, Port: uint16(500), HostPort: uint16(65000)},
+		{Proto: types.UDP, Port: 400, HostPort: 54000},
+		{Proto: types.TCP, Port: 500, HostPort: 65000},
+		{Proto: types.SCTP, Port: 500, HostPort: 65000},
 	}
 
 	sbOptions := make(map[string]interface{})
